refactors to match torch.nn interface
