+++ src/panndas/data/utils.py
@@ -1,0 +1,6 @@
+import numpy as np
+import pandas as pd
+
+
+def eye(index):
+    return pd.DataFrame(np.eye(len(index)), index=index, columns=index)
