+++ src/panndas/nn/__init__.py
@@ -1,11 +1,20 @@
-from .functional import LayerMaxNorm, ReLU
-from .modules import AdditiveSkip, Linear, LinearAttention, Module
-from .modules import Sequential, SoftmaxAttention
+from .modules import (
+    AdditiveSkip,
+    Identity,
+    LayerMaxNorm,
+    Linear,
+    LinearAttention,
+    Module,
+    ReLU,
+    Sequential,
+    SoftmaxAttention,
+)
 
 __all__ = [
     "Module",
     "Sequential",
     "AdditiveSkip",
+    "Identity",
     "Linear",
     "LinearAttention",
     "SoftmaxAttention",
